refactored Form class into 1 parent and 2 child classes
